Update gs_usb.h to reflect the latest changes to the Linux kernel driver.
The change to gs_usb is made to match the changes present in the 5.19 release of the Linux kernel.  The gs_host_frame is now variable size based on whether the data being transferred is CANFD or Classic CAN.  The benefit is that classic CAN frames will not incur the overhead associated with the longer CANFD frames (8-bytes of data vs. 64-bytes).  The CAN handler will need to be updated to reflect the "classic_can" stuct member.