Add multichannel CANFD functionality
Few notes and this I need help with during the review:

Full CANFD support - To implement I needed to do the following to maintain good struct hygine:
- I changed the F0/F4 (bxCAN) handler to use the CubeMX HAL.  I know there is a lot of passion around this.  I was struggling to get the non-HAL and HAL CAN code playing nice but every solution was a mess.  There is a small increase in ROM usage and no increase in RAM.  The cleanliness and portability makes up for it though.
- Added true support for multichannel devices.  The support was there in the gs_usb frame but not in the handler.  I added functions to handle sorting incoming and outgoing CAN messages to the right channel

Notes:
All devices that use the FDCAN architecture (G0/G4/H7) will be able to use this code.
The code is confirmed working with the CANFD changes as of the 5.19 Linux kernel

ISSUES I NEED HELP WITH:
The size of the gs_host_frame has grown from 24 bytes to 80 bytes!  Right now I left the CAN_QUEUE_SIZE as 64 so the code will not compile on any of the devices with 6K of RAM.  I need advice on what do to.  There are a few options but they all have drawbacks:
- Add #defines to the gs_usb.h file to swap the gs_host_frame between the non-FD and FD version - Rquires editing the gs_usb.h. <- This requires the least amount of effort
- Somewhere in the code replace the gs_host_frame with a precompiler variable and map accordingly.  <- Requires touching every file that uses the host frame struct
- Change the CAN_QUEUE_SIZE to something smaller than 64.  In my experince with this driver if you fill up that queue you've got bigger problems then dropping frames. <- This still wastes a ton of memory on non
- Figure out a way to have variable size members on the queue? <- Sure it can be done but ????
- Any combination of the above